+++ lib/empathetech_flutter_ui.dart
@@ -92,11 +92,14 @@ export 'src/widgets/responsive_design/swap_widget.dart';
 export 'src/widgets/responsive_design/transitions.dart';
 
 // Screen reader support
+export 'src/widgets/screen_reader_support/links/elevated_link.dart';
+export 'src/widgets/screen_reader_support/links/icon_link.dart';
+export 'src/widgets/screen_reader_support/links/inline_link.dart';
+export 'src/widgets/screen_reader_support/links/image_link.dart';
+export 'src/widgets/screen_reader_support/links/link_widget.dart';
+export 'src/widgets/screen_reader_support/links/link.dart';
+
 export 'src/widgets/screen_reader_support/image.dart';
-export 'src/widgets/screen_reader_support/inline_link.dart';
-export 'src/widgets/screen_reader_support/link_image_provider.dart';
-export 'src/widgets/screen_reader_support/link_widget.dart';
-export 'src/widgets/screen_reader_support/link.dart';
 export 'src/widgets/screen_reader_support/plain_text.dart';
 export 'src/widgets/screen_reader_support/rich_text.dart';
